fix componente cars, home renderiza el componente cars para modularizar, se elimino la carpeta css
