+++ client/src/Components/Card/Card.module.css
@@ -11,3 +11,9 @@
 .enlace{
     text-decoration: none;
 }
+.cardFlexContent{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
