+++ client/src/Components/Home/Home.module.css
@@ -1,0 +1,15 @@
+.cards{
+    display: grid;
+    grid-template-columns: repeat(3,1fr);
+    grid-gap: 10px;
+
+}
+
+
+.linkstyle{
+    text-decoration: none;
+}
+
+
+
+
